+++ packages/core/cairo_alpha/src/generate/erc1155.ts
@@ -5,6 +5,8 @@ import { infoOptions } from '../set-info';
 import { upgradeableOptions } from '../set-upgradeable';
 import type { RoyaltyInfoSubset } from '../set-royalty-info';
 import { resolveRoyaltyOptionsSubset } from '../set-royalty-info';
+import type { MacrosSubset } from '../set-macros';
+import { resolveMacrosOptions } from '../set-macros';
 import { generateAlternatives } from './alternatives';
 
 const booleans = [true, false];
@@ -12,10 +14,10 @@ const booleans = [true, false];
 type GeneratorOptions = {
   access: AccessSubset;
   royaltyInfo: RoyaltyInfoSubset;
+  macros: MacrosSubset;
 };
 
 function prepareBlueprint(opts: GeneratorOptions) {
-  const royaltyInfo = resolveRoyaltyOptionsSubset(opts.royaltyInfo);
   return {
     name: ['MyToken'],
     baseUri: ['https://example.com/'],
@@ -24,9 +26,10 @@ function prepareBlueprint(opts: GeneratorOptions) {
     mintable: booleans,
     updatableUri: booleans,
     upgradeable: upgradeableOptions,
-    royaltyInfo,
+    royaltyInfo: resolveRoyaltyOptionsSubset(opts.royaltyInfo),
     access: resolveAccessControlOptions(opts.access),
     info: infoOptions,
+    macros: resolveMacrosOptions(opts.macros),
   };
 }