+++ packages/core/cairo_alpha/src/generate/erc20.ts
@@ -4,10 +4,13 @@ import { resolveAccessControlOptions } from '../set-access-control';
 import { infoOptions } from '../set-info';
 import { upgradeableOptions } from '../set-upgradeable';
 import { generateAlternatives } from './alternatives';
+import type { MacrosSubset } from '../set-macros';
+import { resolveMacrosOptions } from '../set-macros';
 
 const booleans = [true, false];
 
 type GeneratorOptions = {
+  macros: MacrosSubset;
   access: AccessSubset;
 };
 
@@ -26,6 +29,7 @@ function prepareBlueprint(opts: GeneratorOptions) {
     access: resolveAccessControlOptions(opts.access),
     upgradeable: upgradeableOptions,
     info: infoOptions,
+    macros: resolveMacrosOptions(opts.macros),
   };
 }