+++ packages/core/cairo_alpha/src/index.ts
@@ -11,11 +11,13 @@ export type { Account } from './account';
 export type { Upgradeable } from './set-upgradeable';
 export type { Info } from './set-info';
 export type { RoyaltyInfoOptions } from './set-royalty-info';
+export type { MacrosOptions } from './set-macros';
 
 export { premintPattern } from './erc20';
 
 export { defaults as infoDefaults } from './set-info';
 export { defaults as royaltyInfoDefaults } from './set-royalty-info';
+export { defaults as macrosDefaults } from './set-macros';
 
 export type { OptionsErrorMessages } from './error';
 export { OptionsError } from './error';