+++ packages/core/cairo_alpha/src/print.ts
@@ -20,6 +20,23 @@ const STANDALONE_IMPORTS_GROUP = 'Standalone Imports';
 const MAX_USE_CLAUSE_LINE_LENGTH = 90;
 const TAB = '\t';
 
+function printWithComponentsDirective(contract: Contract): Lines[] {
+  if (!contract.macros.withComponents || contract.components.length === 0) {
+    return [];
+  }
+  const componentsStr = contract.components
+    .map(c => c.name)
+    .map(name => {
+      const idx = name.indexOf('Component');
+      if (idx === -1) {
+        throw new Error(`Component name "${name}" must contain "Component" substring`);
+      }
+      return name.substring(0, idx);
+    })
+    .join(', ');
+  return [`#[with_components(${componentsStr})]`];
+}
+
 export function printContract(contract: Contract): string {
   const contractAttribute = contract.account ? '#[starknet::contract(account)]' : '#[starknet::contract]';
   return formatLines(
@@ -32,6 +49,7 @@ export function printContract(contract: Contract): string {
       [
         ...printDocumentations(contract.documentations),
         `${contractAttribute}`,
+        ...printWithComponentsDirective(contract),
         `mod ${contract.name} {`,
         spaceBetween(
           printUseClauses(contract),
@@ -156,18 +174,17 @@ function printConstants(constants: Variable[]): Lines[] {
 }
 
 function printComponentDeclarations(contract: Contract): Lines[] {
-  const lines = [];
-  for (const component of contract.components) {
-    lines.push(
-      `component!(path: ${component.name}, storage: ${component.substorage.name}, event: ${component.event.name});`,
-    );
+  if (contract.macros.withComponents) {
+    return [];
   }
-  return lines;
+  return contract.components.map(
+    c => `component!(path: ${c.name}, storage: ${c.substorage.name}, event: ${c.event.name});`,
+  );
 }
 
 function printImpls(contract: Contract): Lines[] {
-  const impls = contract.components.flatMap(c => c.impls);
-
+  const withComponents = contract.macros.withComponents;
+  const impls = contract.components.flatMap(c => (withComponents ? c.impls.filter(i => i.embed) : c.impls));
   // group by section
   const grouped = impls.reduce((result: { [section: string]: Impl[] }, current: Impl) => {
     // default section depends on embed
@@ -202,35 +219,28 @@ function printImpl(impl: Impl): Lines[] {
 }
 
 function printStorage(contract: Contract): (string | string[])[] {
-  const lines = [];
-  // storage is required regardless of whether there are components
-  lines.push('#[storage]');
-  lines.push('struct Storage {');
+  if (contract.macros.withComponents || contract.components.length === 0) {
+    // storage is required regardless of whether there are components
+    return ['#[storage]', 'struct Storage {}'];
+  }
   const storageLines = [];
   for (const component of contract.components) {
     storageLines.push(`#[substorage(v0)]`);
     storageLines.push(`${component.substorage.name}: ${component.substorage.type},`);
   }
-  lines.push(storageLines);
-  lines.push('}');
-  return lines;
+  return ['#[storage]', 'struct Storage {', storageLines, '}'];
 }
 
 function printEvents(contract: Contract): (string | string[])[] {
-  const lines = [];
-  if (contract.components.length > 0) {
-    lines.push('#[event]');
-    lines.push('#[derive(Drop, starknet::Event)]');
-    lines.push('enum Event {');
-    const eventLines = [];
-    for (const component of contract.components) {
-      eventLines.push('#[flat]');
-      eventLines.push(`${component.event.name}: ${component.event.type},`);
-    }
-    lines.push(eventLines);
-    lines.push('}');
+  if (contract.macros.withComponents || contract.components.length === 0) {
+    return [];
   }
-  return lines;
+  const eventLines = [];
+  for (const component of contract.components) {
+    eventLines.push('#[flat]');
+    eventLines.push(`${component.event.name}: ${component.event.type},`);
+  }
+  return ['#[event]', '#[derive(Drop, starknet::Event)]', 'enum Event {', eventLines, '}'];
 }
 
 function printImplementedTraits(contract: Contract): Lines[] {