+++ packages/ui/api/ai-assistant/function-definitions/cairo-alpha.ts
@@ -27,6 +27,7 @@ export const cairoAlphaERC20AIFunctionDefinition = {
         'access',
         'upgradeable',
         'info',
+        'macros',
         'appName',
         'appVersion',
       ]),
@@ -63,6 +64,7 @@ export const cairoAlphaERC721AIFunctionDefinition = {
         'mintable',
         'upgradeable',
         'info',
+        'macros',
         'royaltyInfo',
         'appName',
         'appVersion',
@@ -96,6 +98,7 @@ export const cairoAlphaERC1155AIFunctionDefinition = {
         'access',
         'upgradeable',
         'info',
+        'macros',
         'royaltyInfo',
       ]),
       baseUri: {
@@ -122,6 +125,7 @@ export const cairoAlphaGovernorAIFunctionDefinition = {
         'name',
         'upgradeable',
         'info',
+        'macros',
         'appName',
         'appVersion',
       ]),
@@ -187,7 +191,7 @@ export const cairoAlphaVestingAIFunctionDefinition = {
   parameters: {
     type: 'object',
     properties: {
-      ...addFunctionPropertiesFrom(cairoAlphaSharedFunctionDefinition, ['name', 'info']),
+      ...addFunctionPropertiesFrom(cairoAlphaSharedFunctionDefinition, ['name', 'info', 'macros']),
       startDate: {
         type: 'string',
         description: cairoVestingDescriptions.startDate,
@@ -217,7 +221,7 @@ export const cairoAlphaAccountAIFunctionDefinition = {
   parameters: {
     type: 'object',
     properties: {
-      ...addFunctionPropertiesFrom(cairoAlphaSharedFunctionDefinition, ['name', 'upgradeable', 'info']),
+      ...addFunctionPropertiesFrom(cairoAlphaSharedFunctionDefinition, ['name', 'upgradeable', 'info', 'macros']),
       type: {
         type: 'string',
         enum: ['stark', 'eth'],
@@ -245,7 +249,7 @@ export const cairoAlphaMultisigAIFunctionDefinition = {
   parameters: {
     type: 'object',
     properties: {
-      ...addFunctionPropertiesFrom(cairoAlphaSharedFunctionDefinition, ['name', 'upgradeable', 'info']),
+      ...addFunctionPropertiesFrom(cairoAlphaSharedFunctionDefinition, ['name', 'upgradeable', 'info', 'macros']),
       quorum: {
         type: 'string',
         description: cairoMultisigDescriptions.quorum,
@@ -268,6 +272,7 @@ export const cairoAlphaCustomAIFunctionDefinition = {
         'access',
         'upgradeable',
         'info',
+        'macros',
       ]),
     },
     required: ['name'],