+++ packages/ui/src/cairo_alpha/ERC1155Controls.svelte
@@ -2,18 +2,20 @@
   import HelpTooltip from '../common/HelpTooltip.svelte';
 
   import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo-alpha';
-  import { erc1155, infoDefaults } from '@openzeppelin/wizard-cairo-alpha';
+  import { erc1155, infoDefaults, macrosDefaults } from '@openzeppelin/wizard-cairo-alpha';
 
   import AccessControlSection from './AccessControlSection.svelte';
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import RoyaltyInfoSection from './RoyaltyInfoSection.svelte';
   import InfoSection from './InfoSection.svelte';
+  import MacrosSection from './MacrosSection.svelte';
 
   export let opts: Required<KindedOptions['ERC1155']> = {
     kind: 'ERC1155',
     ...erc1155.defaults,
     royaltyInfo: { ...erc1155.defaults.royaltyInfo }, // copy fields
     info: { ...infoDefaults }, // create new object since Info is nested
+    macros: { ...macrosDefaults }, // create new object since MacrosOptions is nested
     access: { ...erc1155.defaults.access }, // create new object since Access is nested
   };
 
@@ -84,5 +86,7 @@
   required={requireAccessControl}
   {errors}
 />
+
+<MacrosSection bind:macros={opts.macros} />
 
 <InfoSection bind:info={opts.info} />