+++ packages/ui/src/cairo_alpha/ERC721Controls.svelte
@@ -2,12 +2,13 @@
   import HelpTooltip from '../common/HelpTooltip.svelte';
 
   import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo-alpha';
-  import { erc721, infoDefaults } from '@openzeppelin/wizard-cairo-alpha';
+  import { erc721, infoDefaults, macrosDefaults } from '@openzeppelin/wizard-cairo-alpha';
 
   import AccessControlSection from './AccessControlSection.svelte';
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import RoyaltyInfoSection from './RoyaltyInfoSection.svelte';
   import InfoSection from './InfoSection.svelte';
+  import MacrosSection from './MacrosSection.svelte';
   import ExpandableCheckbox from '../common/ExpandableCheckbox.svelte';
   import { error } from '../common/error-tooltip';
 
@@ -15,7 +16,8 @@
     kind: 'ERC721',
     ...erc721.defaults,
     royaltyInfo: { ...erc721.defaults.royaltyInfo }, // copy fields
-    info: { ...infoDefaults }, // create new object since Info is nested,
+    info: { ...infoDefaults }, // create new object since Info is nested
+    macros: { ...macrosDefaults }, // create new object since MacrosOptions is nested
     access: { ...erc721.defaults.access }, // create new object since Access is nested
   };
 
@@ -120,5 +122,7 @@
   required={requireAccessControl}
   {errors}
 />
+
+<MacrosSection bind:macros={opts.macros} />
 
 <InfoSection bind:info={opts.info} />